# api/openapi.yaml
openapi: 3.0.3
info:
  title: Money Manager API
  description: API for managing users, spending, and income
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /users:
    get:
      summary: Get all users
      operationId: getAllUsers
      tags:
        - Users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string

  /users/{id}:
    get:
      summary: Get user by ID
      operationId: getUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      summary: Update user
      operationId: updateUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: User not found
    delete:
      summary: Delete user
      operationId: deleteUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: User not found

  /spending:
    get:
      summary: Get all spending entries
      operationId: getAllSpending
      tags:
        - Spending
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
          description: Filter by user ID
      responses:
        '200':
          description: List of spending entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spending'
    post:
      summary: Create a new spending entry
      operationId: createSpending
      tags:
        - Spending
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpendingRequest'
      responses:
        '200':
          description: Spending created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  message:
                    type: string

  /spending/{id}:
    get:
      summary: Get spending entry by ID
      operationId: getSpending
      tags:
        - Spending
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Spending ID
      responses:
        '200':
          description: Spending entry details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spending'
        '404':
          description: Spending not found
    put:
      summary: Update spending entry
      operationId: updateSpending
      tags:
        - Spending
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Spending ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpendingRequest'
      responses:
        '200':
          description: Spending updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Spending not found
    delete:
      summary: Delete spending entry
      operationId: deleteSpending
      tags:
        - Spending
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Spending ID
      responses:
        '200':
          description: Spending deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Spending not found

  /income:
    get:
      summary: Get all income entries
      operationId: getAllIncome
      tags:
        - Income
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
          description: Filter by user ID
      responses:
        '200':
          description: List of income entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Income'
    post:
      summary: Create a new income entry
      operationId: createIncome
      tags:
        - Income
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncomeRequest'
      responses:
        '200':
          description: Income created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  message:
                    type: string

  /income/{id}:
    get:
      summary: Get income entry by ID
      operationId: getIncome
      tags:
        - Income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Income ID
      responses:
        '200':
          description: Income entry details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Income'
        '404':
          description: Income not found
    put:
      summary: Update income entry
      operationId: updateIncome
      tags:
        - Income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Income ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncomeRequest'
      responses:
        '200':
          description: Income updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Income not found
    delete:
      summary: Delete income entry
      operationId: deleteIncome
      tags:
        - Income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Income ID
      responses:
        '200':
          description: Income deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Income not found

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        username:
          type: string
          description: Username
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - username
        - created_at

    CreateUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username
        password:
          type: string
          format: password
          description: Plain text password (will be hashed before storage)

    UpdateUserRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          description: Username
        password:
          type: string
          format: password
          description: Plain text password (will be hashed before storage, optional - if not provided, existing password is kept)

    Spending:
      type: object
      properties:
        id:
          type: integer
          description: Spending ID
        user_id:
          type: integer
          description: User ID
        category:
          type: string
          description: Spending category
        amount:
          type: number
          format: double
          description: Amount spent
        date:
          type: string
          format: date
          description: Date of spending (YYYY-MM-DD)
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - user_id
        - category
        - amount
        - date
        - created_at

    CreateSpendingRequest:
      type: object
      required:
        - user_id
        - category
        - amount
        - date
      properties:
        user_id:
          type: integer
          description: User ID
        category:
          type: string
          description: Spending category
        amount:
          type: number
          format: double
          description: Amount spent
        date:
          type: string
          format: date
          description: Date of spending (YYYY-MM-DD)

    UpdateSpendingRequest:
      type: object
      required:
        - category
        - amount
        - date
      properties:
        category:
          type: string
          description: Spending category
        amount:
          type: number
          format: double
          description: Amount spent
        date:
          type: string
          format: date
          description: Date of spending (YYYY-MM-DD)

    Income:
      type: object
      properties:
        id:
          type: integer
          description: Income ID
        user_id:
          type: integer
          description: User ID
        category:
          type: string
          description: Income category
        amount:
          type: number
          format: double
          description: Amount earned
        date:
          type: string
          format: date
          description: Date of income (YYYY-MM-DD)
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - user_id
        - category
        - amount
        - date
        - created_at

    CreateIncomeRequest:
      type: object
      required:
        - user_id
        - category
        - amount
        - date
      properties:
        user_id:
          type: integer
          description: User ID
        category:
          type: string
          description: Income category
        amount:
          type: number
          format: double
          description: Amount earned
        date:
          type: string
          format: date
          description: Date of income (YYYY-MM-DD)

    UpdateIncomeRequest:
      type: object
      required:
        - category
        - amount
        - date
      properties:
        category:
          type: string
          description: Income category
        amount:
          type: number
          format: double
          description: Amount earned
        date:
          type: string
          format: date
          description: Date of income (YYYY-MM-DD)